Silence warnings about use of unstable features
Before the feature system was extended to work with standard library
features, we would have needed an #![allow(unstable)] to silence these
warnings. We decided against this at the time, as we wanted to keep
track of anything that wouldn't keep working post-1.0.

The new feature system allows us to keep track of this (via the feature
directive in lib.rs) while at the same time silencing all the warnings.
As 1.0 approaches, we can keep track of unstable feature use and react
accordingly.

Closes #11.