Use IN_NONBLOCK/IN_CLOEXEC by default
It is almost always an error to pass an inotify file descriptor to an exec'd
process, so enable O_CLOEXEC by default.  Also enable nonblock via the argument
to inotify_init1 to save an unneeded fcntl call (this support was added at the
same time as the inotify_init call was replaced with inotify_init1).