Extract FFI module into a dedicated crate
I've already started adding dependencies to the crate that are not
required by the FFI code, and there are more that I'm going to add over
the course of the ongoing cleanup. Saddling any users that are only
interested in the FFI bindings with unneeded dependencies seems
unnecessary, so I believe a second crate makes sense.