Store inotify file descriptor in `Rc` reference
This makes little sense by itself, but it is a first step towards fixing
`WatchDescriptor` equality in the presence of reused file descriptors,
as outlined in issue #60. I decided to split this into its own commit,
as the the many lines this touches would make the diff for the full fix
more confusing than it needs to be.