feat: Add `Inotify::into_event_stream`
This branch allows an `EventStream` to be created from an `Inotify`
but prevents the `Inotify` from closing the file descriptor when
it's dropped. The created `EventStream` will then be responsible for
closing the file descriptor instead.

This allows the `EventStream` to still be used when the `Inotify` that
created it is out of scope, without leaking the file descriptor.

The behaviour of the existing `Inotify::event_stream` function (which
takes `&mut self` rather than `self`, and leaves the `Inotify`
responsible for closing the fd) is unchanged, so this should be
a non-breaking change.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>