refactor: Add guard for closing FDs on drop
This commit improves the close-on-drop behaviour for file descriptors,
by adding a `FdGuard` struct that wraps the file descriptor and is
responsible for closing it when it is dropped. Rather than having an
`Inotify` own an `Arc<RawFd>`, and be responsible for closing it,
`Inotify` now owns an `Arc<FdGuard>`, and the `FdGuard` is responsible
for closing the fd.

This allows us to share the file descriptor between an `Inotify` and
an `EventStream`, and close the file descriptor only when the `Arc` is
dropped (i.e., when both the `Inotify` _and_ the `EventStream` are
dropped). This makes the `Inotify::into_event_stream` I added in #103
unnecessary, so I've gone ahead and deprecated it.