feat: Implement `EventStream` using `mio`
This branch changes the implementation of `Stream` for `EventStream`
to use `mio` and `tokio_reactor` to register interest in events on
the Inotify file descriptor with the reactor, rather than using
`task::current().notify()` to unpark the task in the executor when
polling the fd would block.

Since the reactor is intended to drive all IO resources (such as the
inotify FD) and to be the bridge between tasks in the executor and
events from the OS, using `PollEvented` is a more correct way to
implement the stream.

This branch shouldn't result in any noticeable changes in behaviour,
but will hopefully allow tokio to manage Inotify's IO events more
intelligently.