refactor: Put `EventStream` in its own module
This commit moves `EventStream` to a separate file, since it adds a
lot of related code that's only used for implementing the stream.

Also, since the `EventStream` pulls in a lot of additional
dependencies, and some users may not be using `futures`/`tokio`, I've
added a feature flag, "stream", (on by default) so it can optionally
be disabled if it isn't used.