Add test for DELETE and DELETE_SELF events
Trying to find an issue in my own code, I started thinking that this
crate might not handle deletion events properly, so I wrote tests to
ascertain this. The test pass, so everything is good here. Let's keep
the tests in the repository for the next confused soul and for
regression testing.